move enum to `_utils`
