+++ telegram/_passport/credentials.py
@@ -39,8 +39,8 @@ except ImportError:
 
 from telegram._telegramobject import TelegramObject
 from telegram._utils.argumentparsing import parse_sequence_arg
+from telegram._utils.strings import TextEncoding
 from telegram._utils.types import JSONDict
-from telegram.constants import TextEncoding
 from telegram.error import PassportDecryptionError
 
 if TYPE_CHECKING:
